requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dominate==2.9.1
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
SQLAlchemy==2.0.38
typing_extensions==4.12.2
visitor==0.1.3
Werkzeug==3.1.3
WTForms==3.2.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
Flask-Mail==0.9.1
reportlab==4.1.0
pandas==2.2.0
numpy==1.26.4
